pyproject.toml
[project]
name = "drug-chartor"
version = "1.0.0"
description = "AI-powered drug stability trend chart generator"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "pywin32>=306; platform_system == 'Windows'",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
drug-chartor = "main:main"
